Rename Chritm namespace to CharacterItem
