+++ MikuMikuLibrary/IO/Sections/CharacterItem/ChritmSetSection.cs
@@ -1,0 +1,18 @@
+﻿using System.Text;
+using MikuMikuLibrary.CharacterItem;
+using MikuMikuLibrary.IO.Sections.IO;
+
+namespace MikuMikuLibrary.IO.Sections.CharacterItem
+{
+    [Section("ITEM")]
+    public class AetSetSection : BinaryFileSection<CharacterItemTable>
+    {
+        public override SectionFlags Flags => SectionFlags.None;
+
+        public override Encoding Encoding { get; } = Encoding.GetEncoding("utf-8");
+
+        public AetSetSection(SectionMode mode, CharacterItemTable data = null) : base(mode, data)
+        {
+        }
+    }
+}
